feat(front/redux): add Redux and use it to store loggedUser
This allows to simplify the handling of the logged user.
Logged user is fetched on app load or after login, then stored in redux store.
It can then be retrieved by any component that needs it (e.g.: TopBar which needs to know if user is logged and is admin).

PageGuard now only check if protected children can be accessed thanks to the supplied callback (uses user retrieved from store), and does not fetch loggedUser anymore.

An empty Admin page has been created.

Top Bar has been slightly enhanced, but some work still needs to be done (particularly on menu).