fix(front/sse): add openWhenHidden flag in fetchEventSource to avoid SSE to close and retry
FetchEventSource has the drawback to reuse the same headers on retry, so the auth token will always stay the same on retry.
If a retry occurs once the auth token has expired the trackProgress request would fail with 401.
Using the openWhenHidden flag helps to keep the SSE open no matter what so there is no retry attempts.